requirements.txt
Flask[async]
APScheduler
python-dotenv
Flask-Cors
Flask-Migrate
postgres
psycopg2-binary
Flask-SQLAlchemy
bs4
requests
aiofiles
playwright
slack-sdk
boto3
openai
pillow
asyncio
aiohttp
flasgger
Flask-APScheduler
scikit-learn
pyperclip
google-api-python-client          
flask_migrate       
Flask-SQLAlchemy    
gunicorn
redis
feedparser
pytz
psutil
openpyxl
elementpath
snscrape